How to Use Pagination in Django (Step-by-Step Guide)

+++ ch36/app/models.py
@@ -13,5 +13,5 @@ class Author(models.Model):
 class Book(models.Model):
     title = models.CharField(max_length=100)
     author = models.ForeignKey(Author, on_delete=models.CASCADE)
-    price=models.IntegerField()
-    discount=models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
+    discount = models.IntegerField()
